making dot to have an accreal return type (consistent with CPU)

+++ THCBlas.h
@@ -8,7 +8,7 @@
 THC_API float THCudaBlas_Sdot(THCState *state, long n, float *x, long incx, float *y, long incy);
 THC_API double THCudaBlas_Ddot(THCState *state, long n, double *x, long incx, double *y, long incy);
 #ifdef CUDA_HALF_TENSOR
-THC_API half THCudaBlas_Hdot(THCState *state, long n, half *x, long incx, half *y, long incy);
+THC_API float THCudaBlas_Hdot(THCState *state, long n, half *x, long incx, half *y, long incy);
 #endif
 
 /* Level 2 */
